Apply EXCLUDE simplification to Trino and Spark
